requirements.txt
setuptools
numpy
torch
torchvision
gitpython
wandb
imgaug
pillow
matplotlib
visdom
googledrivedownloader
